requirements.txt
asgiref==3.9.1
click==8.2.1
colorama==0.4.6
crispy-bootstrap5==2025.6
dj-database-url==3.0.1
Django==5.2.4
django-contrib-comments==2.2.0
django-crispy-forms==2.4
django-markdownx==4.0.9
django-recaptcha==4.1.0
gunicorn==23.0.0
h11==0.16.0
Markdown==3.8.2
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
python-decouple==3.8
sqlparse==0.5.3
tzdata==2025.2
uvicorn==0.35.0
whitenoise==6.9.0
